refactor: update rotation functions to use f32 and simplify NdView constructors

+++ src/sfc/math/geo.h
@@ -13,14 +13,14 @@ struct Rot {
     return {1.0f, 0.0f};
   }
 
-  static auto from_rad(f64 angle) -> Rot {
-    const auto c = math::cos(angle);
-    const auto s = math::sin(angle);
+  static auto from_rad(f32 angle) -> Rot {
+    const auto c = math::cosf(angle);
+    const auto s = math::sinf(angle);
     return {f32(c), f32(s)};
   }
 
-  static auto from_deg(f64 deg) -> Rot {
-    const auto a = deg * (math::PI / 180.0);
+  static auto from_deg(f32 deg) -> Rot {
+    const auto a = deg * f32(math::PI / 180.0);
     return Rot::from_rad(a);
   }
